backend/requirements.txt
Flask==3.1.0
flask-restx==1.3.0
flask-cors==4.0.2
python-decouple==3.8
python-dotenv==1.0.1
